+++ frontend/src/components/levelset/fmt.ts
@@ -1,8 +1,24 @@
 export const fmtLevelNumber = (index: number, maxDigits: number) =>
   String(index + 1).padStart(maxDigits, "0");
 
-const fmtLevelTitle = (title: string) =>
-  title.replace(/^[-\s=[\](){}<>~]+|[-\s=[\](){}<>~]+$/g, "");
+const reDecor = /[-\s=[\](){}<>~]+/;
+const reDecorTrim = new RegExp(`^${reDecor.source}|${reDecor.source}$`, "g");
+
+const fmtLevelTitle = (title: string) => title.replace(reDecorTrim, "");
+
+const reFNClean = /[<>:?*|/\\\n]+/g;
+const reFNDecor = /[-\s=[\](){}~.]+/;
+const reFNTrim = new RegExp(`^${reFNDecor.source}|${reFNDecor.source}$`, "g");
+
+export const fmtLevelForFilename = (index: number, title: string) => {
+  let res = String(index + 1);
+
+  const titleFmt = title.replace(reFNClean, ".").replace(reFNTrim, "");
+  if (titleFmt) {
+    res += "." + titleFmt;
+  }
+  return res;
+};
 
 export const fmtLevelShort = (
   index: number,